Move category logic
